+++ src/Test/emc.camus.api.integration.test/Common/RateLimitingIpPartitionTests.cs
@@ -3,7 +3,7 @@ using System.Net;
 using emc.camus.api.integration.test.Fixtures;
 using emc.camus.api.integration.test.Helpers;
 using emc.camus.application.Common;
-using emc.camus.application.RateLimiting;
+using emc.camus.api.Configurations;
 using FluentAssertions;
 
 namespace emc.camus.api.integration.test.Common;