+++ src/Test/emc.camus.api.test/Configurations/RateLimitingSettingsTests.cs
@@ -1,6 +1,5 @@
 using FluentAssertions;
 using emc.camus.api.Configurations;
-using emc.camus.application.RateLimiting;
 
 namespace emc.camus.api.test.Configurations;